Fix allocation reject form
Don't include quota information for reject form

+++ nectar_dashboard/rcallocation/views.py
@@ -326,7 +326,8 @@ class BaseAllocationView(UpdateView):
 
         kwargs = {'form': form}
         # quota
-        kwargs['quota_formsets'] = self.get_quota_formsets()
+        if self.quota_form_class:
+            kwargs['quota_formsets'] = self.get_quota_formsets()
 
         # investigator
         formset_investigator_class = self.get_formset_investigator_class()
@@ -358,7 +359,6 @@ class BaseAllocationView(UpdateView):
         form_class = self.get_form_class()
         form = self.get_form(form_class)
         kwargs = {'form': form}
-        quota_formsets = self.get_quota_formsets()
 
         formset_investigator_class = self.get_formset_investigator_class()
         if formset_investigator_class:
@@ -379,9 +379,12 @@ class BaseAllocationView(UpdateView):
             kwargs['grant_formset'] = self.get_formset(formset_grant_class)
 
         quota_valid = True
-        for service_type, formset in quota_formsets:
-            if not formset.is_valid():
-                quota_valid = False
+        quota_formsets = None
+        if self.quota_form_class:
+            quota_formsets = self.get_quota_formsets()
+            for service_type, formset in quota_formsets:
+                if not formset.is_valid():
+                    quota_valid = False
 
         if quota_valid and all(map(methodcaller('is_valid'), kwargs.values())):
             return self.form_valid(quota_formsets=quota_formsets, **kwargs)
@@ -411,17 +414,18 @@ class BaseAllocationView(UpdateView):
         # quota formsets handled slightly differently as we want to
         # drop objects if requested_quota == 0
         # Default quotas are zero so requesting 0 is not needed
-        for service_type, quota_formset in quota_formsets:
-            quotas = quota_formset.save(commit=False)
-            for obj in quota_formset.deleted_objects:
-                obj.delete()
-            for quota in quotas:
-                if quota.requested_quota > 0:
-                    quota.allocation = self.object
-                    quota.save()
-                else:
-                    if quota.id:
-                        quota.delete()
+        if quota_formsets:
+            for service_type, quota_formset in quota_formsets:
+                quotas = quota_formset.save(commit=False)
+                for obj in quota_formset.deleted_objects:
+                    obj.delete()
+                for quota in quotas:
+                    if quota.requested_quota > 0:
+                        quota.allocation = self.object
+                        quota.save()
+                    else:
+                        if quota.id:
+                            quota.delete()
 
         formsets = [investigator_formset, institution_formset,
                     publication_formset, grant_formset]
